[#33331] Do not force `integer(kind=8)` in subroutines interfaced to 3rd party libraries

See merge request codeaster/src!652

+++ bibfor/include/asterfort/amumpi.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2021 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -22,10 +22,10 @@
 !
 interface
     subroutine amumpi(option, lquali, ldist, kxmps, type, lmhpc, lbloc)
-        integer :: option
+        integer(kind=4) :: option
         aster_logical :: lquali
         aster_logical :: ldist
-        integer :: kxmps
+        integer(kind=8) :: kxmps
         character(len=1) :: type
         aster_logical :: lmhpc
         aster_logical :: lbloc
