+++ bibfor/include/asterfort/asmpi_comm_vect.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -23,13 +23,13 @@
      &,vr,vc,sci,sci4,scr,scc)
               character(len=*), intent(in) :: optmpi
               character(len=*), intent(in) :: typsca
-              integer ,optional, intent(in) :: nbval
-              integer ,optional, intent(in) :: bcrank
-              integer ,optional, intent(inout) :: vi(*)
+              integer(kind=8) ,optional, intent(in) :: nbval
+              integer(kind=8) ,optional, intent(in) :: bcrank
+              integer(kind=8) ,optional, intent(inout) :: vi(*)
               integer(kind=4) ,optional, intent(inout) :: vi4(*)
               real(kind=8) ,optional, intent(inout) :: vr(*)
               complex(kind=8) ,optional, intent(inout) :: vc(*)
-              integer ,optional, intent(inout) :: sci
+              integer(kind=8) ,optional, intent(inout) :: sci
               integer(kind=4) ,optional, intent(inout) :: sci4
               real(kind=8) ,optional, intent(inout) :: scr
               complex(kind=8) ,optional, intent(inout) :: scc
