+++ bibfor/include/asterfort/codent.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2022 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -20,7 +20,7 @@
 !
 interface
     subroutine codent(entier, cadre, chaine, kstop)
-        integer, intent(in) :: entier
+        integer(kind=8), intent(in) :: entier
         character(len=*), intent(in) :: cadre
         character(len=*), intent(out) :: chaine
         character(len=*), optional :: kstop
