+++ bibfor/include/asterfort/jelira.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -22,7 +22,7 @@ interface
     subroutine jelira(nomlu, catr, ival, cval)
         character(len=*), intent(in) :: nomlu
         character(len=*), intent(in) :: catr
-        integer, intent(out), optional  :: ival
+        integer(kind=8), intent(out), optional  :: ival
         character(len=*), intent(out), optional :: cval
     end subroutine jelira
 end interface
